requirements.txt
finnhub-python==2.4.16
loguru==0.6.0
fastavro==1.7.3
fastparquet==2023.2.0
click==8.1.3
confluent_kafka==2.0.2